move timeutils to the time module
