+++ pypy/module/time/timeutils.py
